Moving readResult to api-support.ts (tidy up)
